role model start

+++ src/api/users/users_handlers.py
@@ -9,7 +9,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from src.api.users.actions.auth import get_current_user_from_token
 from src.core.database.models import User
-from src.api.users.actions.user import _create_new_user, _delete_user, _get_user_by_id, _update_user
+from src.api.users.actions.user import _create_new_user, _delete_user, _get_user_by_id, _update_user, check_user_permissions
 from src.api.users.schemas import DeleteUserResponse, ShowUser, UpdatedUserRequest, UpdatedUserResponse, UserCreate
 
 from src.core.database.session import get_db
@@ -35,12 +35,22 @@ async def delete_user(
     db: AsyncSession = Depends(get_db),
     current_user: User = Depends(get_current_user_from_token)
 ) -> DeleteUserResponse:
-    delete_user = await _delete_user(user_id, db)
-    if delete_user is None:
+    user_for_deletion = await _get_user_by_id(user_id, db)
+    if user_for_deletion is None:
         raise HTTPException(
             status_code=404, detail=f"User with id {user_id} not found."
         )
-    return DeleteUserResponse(deleted_user_id=delete_user)
+    if not check_user_permissions(
+        target_user=user_for_deletion,
+        current_user=current_user,
+    ):
+        raise HTTPException(status_code=403, detail="Forbidden.")
+    deleted_user_id = await _delete_user(user_id, db)
+    if deleted_user_id is None:
+        raise HTTPException(
+            status_code=404, detail=f"User with id {user_id} not found."
+        )
+    return DeleteUserResponse(deleted_user_id=deleted_user_id)
 
 
 @router_users.get("/", response_model=ShowUser)
@@ -68,19 +78,25 @@ async def update_user_by_id(user_id: int,
 
     if updated_user_params == {}:
         raise HTTPException(
-            status_code=404, detail="Выберите хотя бы один параметр для обновления")
+            status_code=422, detail="Выберите хотя бы один параметр для обновления")
 
-    user = await _get_user_by_id(user_id, db)
+    user_for_update = await _get_user_by_id(user_id, db)
+    if user_id != current_user.user_id:
+        if not check_user_permissions(
+            target_user=user_for_update, current_user=current_user
+        ):
+            raise HTTPException(status_code=403, detail="Forbidden.")
 
-    if user is None:
+    if user_for_update is None:
         raise HTTPException(
             status_code=404, detail=f"User with id {user_id} not found."
         )
 
     try:
-        updated_user_id = await _update_user(updated_user_params=updated_user_params, session=db, user_id=user_id),
+        updated_user_id = await _update_user(
+            updated_user_params=updated_user_params, session=db, user_id=user_id
+        )
     except IntegrityError as err:
         logger.error(err)
-        raise HTTPException(
-            status_code=503, detail=f"Пользователь с e-mail: {body.email} уже существует")
-    return UpdatedUserResponse(updated_user_id=updated_user_id[0])
+        raise HTTPException(status_code=503, detail=f"Database error: {err}")
+    return UpdatedUserResponse(updated_user_id=updated_user_id)
